requirements.txt
Flask==3.1.3
Flask-Bcrypt==1.0.1
Authlib==1.6.9
requests==2.33.1
feedparser==6.0.12
akshare==1.18.49
python-dotenv==1.2.2
google-genai
